refactor(di): modularize DI container into separate files
- Split monolithic di.ts into backend/di/ directory with repositories.ts, services.ts, routes.ts, and index.ts
- repositories.ts: All repository factory functions
- services.ts: All service factory functions
- routes.ts: Route factories, helpers, and validation logic
- index.ts: Barrel export for clean imports
- Update server.ts import to use new modular structure
- Fix middleware/index.ts to export validateBody instead of validationMiddleware
- Improve maintainability: each file ~50-100 lines, focused responsibilities
- Matches codebase patterns (e.g., backend/repositories/index.ts)
