+++ backend/di/index.ts
@@ -1,0 +1,4 @@
+// Barrel export for DI container
+export * from "./repositories.ts";
+export * from "./services.ts";
+export * from "./routes.ts";
