+++ backend/middleware/index.ts
@@ -1,3 +1,3 @@
 export { authMiddleware } from "./auth.ts";
-export { validationMiddleware } from "./validation.ts";
+export { validateBody } from "./validation.ts";
 export { errorMiddleware } from "./error.ts";
